django/requirements.txt
# at least
coderedcms>=1.0
psycopg2-binary>=2.9
gunicorn>=20.1

#last version
django-getenv
